research_agent/requirements.txt
python-dotenv
langchain
langchain_community
langchain-openai
openai
requests
duckduckgo-search
ipykernel
beautifulsoup4
